requirements.txt
streamlit== 1.51.0
pandas== 2.3.3
python-dotenv == 1.2.1 
pytz == 2025.2
supabase== 2.24.0
requests== 2.32.5
PyGithub== 2.8.1 
matplotlib== 3.10.7
typing_extensions== 4.15.0
langchain                     
langchain-core                
langchain-community          
langchain-groq                
langchain-text-splitters     
streamlit                     
python-dotenv                 
setuptools                    
langchain_google_genai
dateparser
setuptools